feat(common): 新增了Socket常量
